requirements.txt
python-osc==1.8.0
openvr==1.16.802
pyyaml==6.0
pyglm==2.5.7